+++ client/src/app/projects/[id]/project-and-background-community/layout.tsx
@@ -12,8 +12,7 @@ import {
 import getQueryClient from "@/app/getQueryClient";
 import { ProjectsDetailPageProps } from "@/app/projects/[id]/page";
 
-import NavigationSidebar, { NavigationSidebarProps } from "@/containers/navigation/sidebar";
-import PCBSidebarItem from "@/containers/projects/detail/sidebar/pcb-item";
+import PcbSidebar from "@/containers/projects/detail/sidebar/pcb";
 import Sidebar from "@/containers/sidebar";
 
 interface ProjectsDetailPCBLayoutProps extends ProjectsDetailPageProps, PropsWithChildren {}
@@ -25,12 +24,7 @@ export async function generateMetadata({ params }: ProjectsDetailPageProps): Pro
   };
 }
 
-export default async function ProjectsDetailPCBLayout({
-  children,
-  params,
-}: ProjectsDetailPCBLayoutProps) {
-  const { id } = params;
-
+export default async function ProjectsDetailPCBLayout({ children }: ProjectsDetailPCBLayoutProps) {
   const CATEGORIES = await getPcbCategories({
     sort: "display_order:asc",
   });
@@ -45,42 +39,13 @@ export default async function ProjectsDetailPCBLayout({
 
   const dehydratedState = dehydrate(queryClient);
 
-  const items = [
-    {
-      href: `/projects/${id}/project-and-background-community`,
-      label: "Overview",
-      children: <span className="text-lg">Overview</span>,
-    },
-    ...(CATEGORIES?.data || [])
-      ?.sort((a, b) => {
-        if (a?.attributes?.display_order && b?.attributes?.display_order) {
-          return +a.attributes.display_order - +b.attributes.display_order;
-        }
-
-        return 0;
-      })
-      ?.map(({ id: categoryId, attributes }) => {
-        return {
-          href: `/projects/${id}/project-and-background-community/${categoryId}`,
-          label: attributes?.title ?? "",
-          children: (
-            <>
-              {typeof categoryId !== "undefined" && <PCBSidebarItem categoryId={categoryId} />}
-
-              <span>{attributes?.title}</span>
-            </>
-          ),
-        };
-      }),
-  ] satisfies NavigationSidebarProps["items"];
-
   return (
     <Hydrate state={dehydratedState}>
       <section className="flex grow flex-col space-y-5">
         <div className="grid grid-cols-12 gap-20">
           <div className="col-span-4">
             <Sidebar>
-              <NavigationSidebar items={items} />
+              <PcbSidebar />
             </Sidebar>
           </div>
           <div className="col-span-8">{children}</div>
