+++ client/src/containers/projects/detail/sidebar/contextual-risk/index.tsx
@@ -1,0 +1,49 @@
+"use client";
+
+import { useParams } from "next/navigation";
+
+import { useGetContextualRiskCategories } from "@/types/generated/contextual-risk-category";
+
+import NavigationSidebar, { NavigationSidebarProps } from "@/containers/navigation/sidebar";
+import ContextualRiskSidebarItem from "@/containers/projects/detail/sidebar/contextual-risk/item";
+
+export default function ContextualRiskSidebar() {
+  const { id } = useParams();
+
+  const { data: categoriesData } = useGetContextualRiskCategories({
+    sort: "display_order:asc",
+  });
+
+  const items = [
+    {
+      href: `/projects/${id}/contextual-risk`,
+      label: "Overview",
+      children: <span className="text-lg">Overview</span>,
+    },
+    ...(categoriesData?.data || [])
+      ?.sort((a, b) => {
+        if (a?.attributes?.display_order && b?.attributes?.display_order) {
+          return +a.attributes.display_order - +b.attributes.display_order;
+        }
+
+        return 0;
+      })
+      ?.map(({ id: categoryId, attributes }) => {
+        return {
+          href: `/projects/${id}/contextual-risk/${categoryId}`,
+          label: attributes?.title ?? "",
+          children: (
+            <>
+              {typeof categoryId !== "undefined" && (
+                <ContextualRiskSidebarItem categoryId={categoryId} />
+              )}
+
+              <span>{attributes?.title}</span>
+            </>
+          ),
+        };
+      }),
+  ] satisfies NavigationSidebarProps["items"];
+
+  return <NavigationSidebar items={items} />;
+}
